improve validation
The `min`, `max`, and `eq` functions can each test interfaces that have `length` parameter, such as `array`/`slice`/`map`/`chan`/`string`. Thus, calling the `min` function in the context of these interfaces compares the minimum length of these interfaces, and the rules such as `minLen`, `maxLen`, and `len` are synonyms for `min`, `max`, and `eq`, respectively.